test(tensor): keep optimization tests module-local
